Change ll to int

+++ Sqrt Decomposition/sqrtDecomposition1.cpp
@@ -6,39 +6,39 @@ Sqrt Decomposition
 template <typename Type>
 struct sqrt_dec
 {
-	ll len;
+	int len;
 	vector<Type> a, bl;
-	void build(ll n, Type *ar)
+	void build(int n, Type *ar)
 	{
 		len = sqrt(n);
 		bl.resize(n / len + 1);
 		a.push_back(0);
-		for (ll i = 1; i <= n; i++)
+		for (int i = 1; i <= n; i++)
 		{
 			a.push_back(ar[i]);
 			bl[i / len] += ar[i];
 		}
 	}
-	void upd(ll pos, Type k)
+	void upd(int pos, Type k)
 	{
 		a[pos] += k;
 		bl[pos / len] += k;
 	}
-	Type sum(ll l, ll r)
+	Type sum(int l, int r)
 	{
 		Type ans = 0;
-		ll cl = l / len, cr = r / len;
+		int cl = l / len, cr = r / len;
 		if (cl == cr)
 		{
-			for (ll i = l; i <= r; i++)
+			for (int i = l; i <= r; i++)
 				ans += a[i];
 			return ans;
 		}
-		for (ll i = l; i < (cl + 1) * len; i++)
+		for (int i = l; i < (cl + 1) * len; i++)
 			ans += a[i];
-		for (ll i = cl + 1; i < cr; i++)
+		for (int i = cl + 1; i < cr; i++)
 			ans += bl[i];
-		for (ll i = cr * len; i <= r; i++)
+		for (int i = cr * len; i <= r; i++)
 			ans += a[i];
 		return ans;
 	}
